+++ com.unity.netcode.gameobjects/Editor/CodeGen/RuntimeAccessModifiersILPP.cs
@@ -97,17 +97,23 @@ namespace Unity.Netcode.Editor.CodeGen
         {
             foreach (var fieldDefinition in typeDefinition.Fields)
             {
+#pragma warning disable CS0618 // Type or member is obsolete
                 if (fieldDefinition.Name == nameof(NetworkManager.__rpc_func_table))
+#pragma warning restore CS0618 // Type or member is obsolete
                 {
                     fieldDefinition.IsPublic = true;
                 }
 
+#pragma warning disable CS0618 // Type or member is obsolete
                 if (fieldDefinition.Name == nameof(NetworkManager.RpcReceiveHandler))
+#pragma warning restore CS0618 // Type or member is obsolete
                 {
                     fieldDefinition.IsPublic = true;
                 }
 
+#pragma warning disable CS0618 // Type or member is obsolete
                 if (fieldDefinition.Name == nameof(NetworkManager.__rpc_name_table))
+#pragma warning restore CS0618 // Type or member is obsolete
                 {
                     fieldDefinition.IsPublic = true;
                 }
@@ -115,7 +121,9 @@ namespace Unity.Netcode.Editor.CodeGen
 
             foreach (var nestedTypeDefinition in typeDefinition.NestedTypes)
             {
+#pragma warning disable CS0618 // Type or member is obsolete
                 if (nestedTypeDefinition.Name == nameof(NetworkManager.RpcReceiveHandler))
+#pragma warning restore CS0618 // Type or member is obsolete
                 {
                     nestedTypeDefinition.IsNestedPublic = true;
                 }