+++ com.unity.netcode.gameobjects/Runtime/Transports/UTP/INetworkStreamDriverConstructor.cs
@@ -36,9 +36,8 @@ namespace Unity.Netcode.Transports.UTP
     ///             var settings = transport.GetDefaultNetworkSettings();
     ///             driver = NetworkDriver.Create(new IPCNetworkInterface(), settings);
     ///
-    ///             driver.RegisterPipelineStage(new NetworkMetricsPipelineStage());
-    ///
     ///             transport.GetDefaultPipelineConfigurations(
+    ///                 ref driver,
     ///                 out var unreliableFragmentedPipelineStages,
     ///                 out var unreliableSequencedFragmentedPipelineStages,
     ///                 out var reliableSequencedPipelineStages);